feat: poetry and docker support
